Connect backend with database - MongoDB Atlas
Connet backend with cloud database MongoDB and due to
kotlinx.serialization not handling bson type, like _id: ObjectId,
created automatically inside the database and we cannot change it for a
simpler format as String, it was removed and replaced with Jackson.

Things that still need to be done: refactore endpoints.