Return from login with User
After a successfull login with Google, the back end gets the Google
email and searches a User object inside database with that email and
returns the founded user to front end, but if it not exists, the
username is empty `""`.

Also, we added the endpoint `/sign-up` that should receive an User object
in with a POST request. The rules are that `username` must not be
empty, `email` must match with Google account's email, and `isModerator`
must be `false`.

All functionalities were not tested yet! Do not merge it until I ask for it!