Feature/PROJCONN-58 auth improvements (#15)
* add new env variable

* add new dependencies

* add Google tokenVerifier

* update auth

* change logging level for mongo driver

* update status route to return auth status

* protect posts endpoint

* add tokenVerifier usage

* Return from login with User

After a successfull login with Google, the back end gets the Google
email and searches a User object inside database with that email and
returns the founded user to front end, but if it not exists, the
username is empty `""`.

Also, we added the endpoint `/sign-up` that should receive an User object
in with a POST request. The rules are that `username` must not be
empty, `email` must match with Google account's email, and `isModerator`
must be `false`.

All functionalities were not tested yet! Do not merge it until I ask for it!

* Add safe insert User in database

* fix redirect

* MongoDB operations implementation

* add PublicUser data class

* add conversion extension functions

* update User model

* minor fixes

* User routes changes

Co-authored-by: Willian Ozako <willian.ozako@gmail.com>